refactor(binder): split type_inference to smaller files (#3244)
